Big refactor of main file
The main difference is the flattened recursion has been reorganized into
a more normal recursive function call, which makes the code slightly
clearer.

The core export task is now achieved with a class instance (though not
exposed to the caller), which clarifies which variables are local and
which are needed at all levels of the call stack. This also lets us
break the big function down into more methods, and separate out a lot of
the micromanagement tasks into functions.

One breaking change is that in the 'options' key of the output file,
'flattenedOpacity' is now 'flattenOpacity' to match the input option
name.